refactor(filters): rename most components and files
