+++ packages/admin-ui-docs/docs/experimental/filter-group.mdx
@@ -9,14 +9,14 @@ path: /filter-group/
 
 ```jsx
 import {
-  experimental_SingleSelectFilter as SingleSelectFilter,
-  experimental_useSingleFilterState as useSingleFilterState,
+  experimental_Filter as Filter,
+  experimental_useFilterState as useFilterState,
   experimental_useFilterGroupState as useFilterGroupState,
   experimental_FilterGroup as FilterGroup,
 } from '@vtex/admin-ui'
 
 function Example() {
-  const filterState = useMultipleFilterState({
+  const filterState = useFilterMultipleState({
     items: [
       { label: 'Full', value: 1, id: '#1' },
       { label: 'Empty', value: 2, id: '#2' },
@@ -26,21 +26,27 @@ function Example() {
     label: 'Status',
   })
 
-  const state = useFilterGroupState([filterState])
+  const state = useFilterGroupState({filterStates: [filterState]})
 
   return (
     <FilterGroup state={state}>
-        <MultiselectFilter state={state} />
+        <FilterMultiple state={filterState} />
     </FilterGroup>
   )
 }
 ```
 
-# Props
+## Props
 
 ## FilterGroup
 
-| Name  | Type              | Description    | Required | Default |
-| ----- | ----------------- | -------------- | -------- | ------- |
-| state | `FilterGroupState` | Required state | ✅       | -       |
-| csx   | `StyleObject`     | Custom styles  | 🚫       | `{}`    |
+| Name  | Type               | Description    | Required | Default |
+| ----- | -----------------  | -------------- | -------- | ------- |
+| state | `FilterGroupState` | Object that manages the component state logic, you can obtain it calling the useFilterGroupState hook | ✅       | -       |
+| csx   | `StyleObject`      | Custom styles  | 🚫       | `{}`    |
+
+## useFilterGroupState
+
+| Name           | Type                           | Description                                                       | Required | Default |
+| -------------- | ------------------------------ | ----------------------------------------------------------------- | -------- | ------- |
+| filterStates          | `FilterState[]`                | Array of filter states, add the state of each child filter    | ✅       | -       |
