+++ packages/admin-ui/src/filters/filter-group.tsx
@@ -6,13 +6,13 @@ import { Flex } from '../components/Flex'
 import { useMessageFormatter } from '../i18n'
 import { messages } from './filter.i18n'
 
-export function FilterGroup(props: FilterBarProps) {
+export function FilterGroup(props: FilterGroupProps) {
   const { children, state, ...rest } = props
 
   const formatMessage = useMessageFormatter(messages.actions)
 
   return (
-    <Flex wrap="wrap" {...rest}>
+    <Flex wrap="wrap" role="group" {...rest}>
       {children}
       <Button onClick={state.onClear} variant="adaptative-dark" size="small">
         {formatMessage('clearAll')}
@@ -21,7 +21,7 @@ export function FilterGroup(props: FilterBarProps) {
   )
 }
 
-export interface FilterBarProps extends SystemComponentProps<{}> {
+export interface FilterGroupProps extends SystemComponentProps<{}> {
   children?: ReactNode
   state: { onClear: () => void }
 }
