+++ packages/admin-ui/src/filters/filter-multiple.state.tsx
@@ -1,0 +1,11 @@
+import type {
+  UseFilterMultipleReturn,
+  UseFilterMultipleStateProps,
+} from './base-filter.state'
+import { useBaseFilterState } from './base-filter.state'
+
+export function useFilterMultipleState(
+  props: UseFilterMultipleStateProps
+): UseFilterMultipleReturn {
+  return useBaseFilterState({ ...props, selectionMode: 'multiple' })
+}
