+++ packages/admin-ui/src/filters/index.tsx
@@ -1,6 +1,6 @@
-export * from './multiselect-filter'
-export * from './single-select-filter'
+export * from './filter-multiple'
+export * from './filter'
 export * from './filter-group'
 export * from './filter-group.state'
-export * from './multiple-filter.state'
-export * from './single-filter.state'
+export * from './filter-multiple.state'
+export * from './filter.state'
