--- packages/admin-ui/src/filters/multiple-filter.state.tsx
@@ -1,11 +1,0 @@
-import type {
-  UseMultipleFilterReturn,
-  UseMultipleFilterStateProps,
-} from './filter.state'
-import { useFilterState } from './filter.state'
-
-export function useMultipleFilterState(
-  props: UseMultipleFilterStateProps
-): UseMultipleFilterReturn {
-  return useFilterState({ ...props, selectionMode: 'multiple' })
-}
